Preparing for pip (#1581)
* put requirements in setup and add entry points
* moving binaries to onmt.bin.*
* add scripts in root for compatibility
* remove requirements install from travis
* fix test_preprocess
* fix scripts
* fix scripts path for docs
