+++ .travis.yml
@@ -14,7 +14,6 @@ addons:
 before_install:
   # Install CPU version of PyTorch.
   - if [[ $TRAVIS_PYTHON_VERSION == 3.5 ]]; then pip install torch==1.2.0 -f https://download.pytorch.org/whl/cpu/torch_stable.html; fi
-  - pip install -r requirements.txt
   - pip install -r requirements.opt.txt
   - python setup.py install   
 env:
