+++ docs/source/options/preprocess.rst
@@ -2,6 +2,6 @@ Preprocess
 ==========
 
 .. argparse::
-    :filename: ../preprocess.py
+    :filename: ../onmt/bin/preprocess.py
     :func: _get_parser
     :prog: preprocess.py
