+++ docs/source/options/train.rst
@@ -2,6 +2,6 @@ Train
 =====
 
 .. argparse::
-    :filename: ../train.py
+    :filename: ../onmt/bin/train.py
     :func: _get_parser
     :prog: train.py
