+++ docs/source/options/translate.rst
@@ -2,6 +2,6 @@ Translate
 =========
 
 .. argparse::
-    :filename: ../translate.py
+    :filename: ../onmt/bin/translate.py
     :func: _get_parser
     :prog: translate.py
