+++ requirements.opt.txt
@@ -5,9 +5,6 @@ librosa
 Pillow
 git+git://github.com/pytorch/audio.git@d92de5b97fc6204db4b1e3ed20c03ac06f5d53f0
 pyrouge
-pyonmttok
 opencv-python
 git+https://github.com/NVIDIA/apex
-flask
-tensorboard>=1.14
 pretrainedmodels
