+++ setup.py
@@ -1,11 +1,29 @@
 #!/usr/bin/env python
+from setuptools import setup, find_packages
 
-from setuptools import setup
 
-setup(name='OpenNMT-py',
-      description='A python implementation of OpenNMT',
-      version='0.9.2',
-
-      packages=['onmt', 'onmt.encoders', 'onmt.modules', 'onmt.tests',
-                'onmt.translate', 'onmt.decoders', 'onmt.inputters',
-                'onmt.models', 'onmt.utils'])
+setup(
+    name='OpenNMT-py',
+    description='A python implementation of OpenNMT',
+    version='0.9.2',
+    packages=find_packages(),
+    install_requires=[
+        "six",
+        "tqdm~=4.30.0",
+        "torch>=1.1",
+        "torchtext==0.4.0",
+        "future",
+        "configargparse",
+        "tensorboard>=1.14",
+        "flask",
+        "pyonmttok",
+    ],
+    entry_points={
+        "console_scripts": [
+            "onmt_server=onmt.bin.server:main",
+            "onmt_train=onmt.bin.train:main",
+            "onmt_translate=onmt.bin.translate:main",
+            "onmt_preprocess=onmt.bin.preprocess:main",
+        ],
+    }
+)
